app/src/main/java/de/anew/activities/tasks/TaskTouchHelper.kt
package de.anew.activities.tasks

import androidx.recyclerview.widget.ItemTouchHelper
import androidx.recyclerview.widget.RecyclerView

class TaskTouchHelper(private val taskAdapter: TaskAdapter) : ItemTouchHelper.SimpleCallback(0, ItemTouchHelper.RIGHT) {

    override fun onMove(recyclerView: RecyclerView, viewHolder: RecyclerView.ViewHolder, target: RecyclerView.ViewHolder) = false

    override fun onSwiped(viewHolder: RecyclerView.ViewHolder, direction: Int) {
        taskAdapter.deleteTaskAt(viewHolder.adapterPosition)
    }

}